update
update

+++ Java/Loon-Lite(PureJava)/LoonLiteCore/src/loon/action/camera/ScreenViewport.java
@@ -23,7 +23,7 @@ package loon.action.camera;
 public class ScreenViewport extends Viewport {
 
 	@Override
-	public void onResize(int width, int height) {
+	public void onResize(float width, float height) {
 		setBounds(0, 0, width, height, 1f, 1f);
 	}
 }
